src/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.4.2
openai==1.2.0
azure-search-documents==11.4.0
azure-identity==1.14.0
